lurek2d::runtime

Module log_messages

Log message identifiers and log-level override helpers. Stable, structured log message identifiers for all engine subsystems.

Each constant provides a short code (e.g. “L001”) used as prefix in log output.

Identifiers grouped by domain: L=lifecycle, A=audio, G=graphics, P=physics, FS=filesystem.

Additional prefixes: AN=animation, EN=ECS, TM=tilemap, SV=save, SC=scene, TH=thread, PF=pathfind.

Extended prefixes: MD=mods, NW=network, PL=pipeline, AT=automation, CP=compute, SR=serial, GU=GUI.

Runtime log level control via set_log_level/get_log_level with atomic override.

log_msg! macro for consistent formatted log output with message lookup.

Codes are stable across versions for log parsing, alerting, and external tool integration.

Constants

A001_MIDI_READ_FAIL

Stable log message identifier.

A002_MIDI_DISABLED

Stable log message identifier.

A003_AUDIO_OUTPUT_UNAVAIL

Stable log message identifier.

A004_AUDIO_PLAY_QUEUED

Stable log message identifier.

AD01_AUDIO_DECODED

Stable log message identifier.

AD02_AUDIO_ERROR

Stable log message identifier.

AN01_ANIM_CTRL_INIT

Stable log message identifier.

AN02_CLIP_ADDED

Stable log message identifier.

AN03_CLIP_NOT_FOUND

Stable log message identifier.

AS01

Stable log message identifier.

AT01

Stable log message identifier.

AT01_SIM_INIT

Stable log message identifier.

AT02

Stable log message identifier.

AT02_SCRIPT_LOAD

Stable log message identifier.

AT03

Stable log message identifier.

AT03_STEP_WARN

Stable log message identifier.

BB01

Stable log message identifier.

BB02

Stable log message identifier.

BB03

Stable log message identifier.

BD01

Stable log message identifier.

BD02

Stable log message identifier.

BD03

Stable log message identifier.

BI01

Stable log message identifier.

BI02

Stable log message identifier.

BI03

Stable log message identifier.

BU01

Stable log message identifier.

BU02

Stable log message identifier.

BU03

Stable log message identifier.

CB01

Stable log message identifier.

CB02

Stable log message identifier.

CH01

Stable log message identifier.

CH02

Stable log message identifier.

CH03

Stable log message identifier.

CH04

Stable log message identifier.

CK01

Stable log message identifier.

CK02

Stable log message identifier.

CK03

Stable log message identifier.

CP01_NDARRAY_ALLOC

Stable log message identifier.

CP02_NDARRAY_LARGE

Stable log message identifier.

CQ01

Stable log message identifier.

CQ02

Stable log message identifier.

CQ03

Stable log message identifier.

CV01

Stable log message identifier.

DF01

Stable log message identifier.

DF01_RIGHT_JOIN_UNIMPL

Stable log message identifier.

DP01

Stable log message identifier.

DP02

Stable log message identifier.

DP03

Stable log message identifier.

EN01_UNIVERSE_INIT

Stable log message identifier.

EN02_ENTITY_SPAWN

Stable log message identifier.

EN03_ENTITY_LOW

Stable log message identifier.

FE01

Stable log message identifier.

FE02

Stable log message identifier.

FE03

Stable log message identifier.

FF01

Stable log message identifier.

FF02

Stable log message identifier.

FF03

Stable log message identifier.

FN01

Stable log message identifier.

FN02

Stable log message identifier.

FS01_GAMEFS_INIT

Stable log message identifier.

FS02_FILE_READ

Stable log message identifier.

FS03_FILE_WRITE

Stable log message identifier.

FS04_PATH_TRAVERSAL

Stable log message identifier.

FS05_VFS_MOUNT

Stable log message identifier.

FX01

Stable log message identifier.

FX02

Stable log message identifier.

G001_FONT_GLYPH_WARN

Stable log message identifier.

G002_SCREENSHOT_ZERO_SIZE

Stable log message identifier.

G003_SCREENSHOT_MAP_FAIL

Stable log message identifier.

G004_SCREENSHOT_RECV_FAIL

Stable log message identifier.

G005_SCREENSHOT_DATA_FAIL

Stable log message identifier.

G006_ATLAS_MAX_SIZE

Stable log message identifier.

GC01

Stable log message identifier.

GC02

Stable log message identifier.

GC03

Stable log message identifier.

GC04

Stable log message identifier.

GD01

Stable log message identifier.

GD02

Stable log message identifier.

GD03

Stable log message identifier.

GP01

Stable log message identifier.

GP02

Stable log message identifier.

GP03

Stable log message identifier.

GR01

Stable log message identifier.

GR02

Stable log message identifier.

GU01_CTX_INIT

Stable log message identifier.

GU02_WIDGET_ADD

Stable log message identifier.

GU03_CTX_RESET

Stable log message identifier.

HP01

Stable log message identifier.

HP02

Stable log message identifier.

HP03

Stable log message identifier.

HX01

Stable log message identifier.

HX02

Stable log message identifier.

IE01

Stable log message identifier.

IE02

Stable log message identifier.

IE03

Stable log message identifier.

IF01

Stable log message identifier.

IF02

Stable log message identifier.

IF03

Stable log message identifier.

IM01_IMAGE_LOADED

Stable log message identifier.

IM02_IMAGE_MISMATCH

Stable log message identifier.

L001_ENGINE_START

Stable log message identifier.

L002_ENGINE_STOP

Stable log message identifier.

L003_GAME_LOADED

Stable log message identifier.

L004_GAME_RESTART

Stable log message identifier.

L005_CONF_LOADED

Stable log message identifier.

L006_SPLASH_SCREEN

Stable log message identifier.

L007_NO_MAIN_LUA

Stable log message identifier.

L010_RENDER_ERROR

Stable log message identifier.

L011_LUA_ERROR

Stable log message identifier.

L012_AUDIO_ERROR

Stable log message identifier.

L013_FS_ERROR

Stable log message identifier.

L014_PHYSICS_ERROR

Stable log message identifier.

L015_RESOURCE_NOT_FOUND

Stable log message identifier.

L016_LUA_VM_INIT_FAIL

Stable log message identifier.

L017_MAIN_LUA_READ_FAIL

Stable log message identifier.

L020_NO_AUDIO_DEVICE

Stable log message identifier.

L021_CLIPBOARD_FAIL

Stable log message identifier.

L022_UNKNOWN_LOG_LEVEL

Stable log message identifier.

L023_GPU_TEX_TOO_SMALL

Stable log message identifier.

L024_SURFACE_LOST

Stable log message identifier.

L030_ASYNC_LOAD_REQUEST

Stable log message identifier.

L031_ASYNC_LOAD_COMPLETE

Stable log message identifier.

L032_BATCH_STATS

Stable log message identifier.

L033_GPU_ADAPTER

Stable log message identifier.

L034_GPU_TEX_DIM

Stable log message identifier.

L035_GPU_INIT

Stable log message identifier.

L036_GAMEPAD_CONNECTED

Stable log message identifier.

L037_GAMEPAD_DISCONNECTED

Stable log message identifier.

L038_GILRS_UNAVAILABLE

Stable log message identifier.

L039_WINDOW_CLOSE

Stable log message identifier.

L040_ICON_LOAD_FAIL

Stable log message identifier.

L041_ICON_CONV_FAIL

Stable log message identifier.

L042_DISPLAY_INDEX_UNAVAIL

Stable log message identifier.

L043_DROP_FILE

Stable log message identifier.

L044_DROP_GAME

Stable log message identifier.

L050_MODULE_DEP_DISABLED

Stable log message identifier.

L051_CONF_READ_ERR

Stable log message identifier.

L052_CONF_PARSE_ERR

Stable log message identifier.

L053_CONF_CALLBACK_ERR

Stable log message identifier.

L060_LUA_CALLBACK_ERROR

Stable log message identifier.

L070_SURFACE_NO_READBACK

Stable log message identifier.

L071_CURSOR_GRAB_FAIL

Stable log message identifier.

L072_CURSOR_GRAB_LOCK_FAIL

Stable log message identifier.

L073_CURSOR_POS_FAIL

Stable log message identifier.

L074_SCREENSHOT_NO_READBACK

Stable log message identifier.

L075_SCREENSHOT_SAVE_FAIL

Stable log message identifier.

L076_SCREENSHOT_ENCODE_FAIL

Stable log message identifier.

L077_DRAG_HOVER

Stable log message identifier.

L078_DRAG_HOVER_CANCEL

Stable log message identifier.

L079_DRAG_DROP_IGNORED

Stable log message identifier.

L080_GAME_DIR

Stable log message identifier.

L081_LOG_FILE

Stable log message identifier.

L082_LOG_FILE_FAIL

Stable log message identifier.

L083_DROP_ARCHIVE

Stable log message identifier.

L084_DROP_ARCHIVE_FAIL

Stable log message identifier.

LA01_API_STUB

Stable log message identifier.

LA02_PIPELINE_CALLBACK_FAIL

Stable log message identifier.

LA03_OPEN_URL_REJECTED

Stable log message identifier.

LA04_CLIPBOARD_WRITE_FAIL

Stable log message identifier.

LA05_CLIPBOARD_UNAVAIL

Stable log message identifier.

LA06_CLIPBOARD_READ_FAIL

Stable log message identifier.

LA07_SCALE_MODE_UNKNOWN

Stable log message identifier.

LA08_PATHFINDING_THREAD_UNIMPL

Stable log message identifier.

LT01

Stable log message identifier.

LT02

Stable log message identifier.

LT03

Stable log message identifier.

LW01_LIGHT_WORLD_INIT

Stable log message identifier.

LW02_LIGHT_ADD

Stable log message identifier.

LW03_LIGHT_MAX

Stable log message identifier.

MD01_MGR_INIT

Stable log message identifier.

MD02_MOD_REG

Stable log message identifier.

MD03_MOD_FAIL

Stable log message identifier.

MD04_ORDER_OK

Stable log message identifier.

MG01

Stable log message identifier.

MG02

Stable log message identifier.

MG03

Stable log message identifier.

MM01_MINIMAP_INIT

Stable log message identifier.

MM02_TERRAIN_REBUILD

Stable log message identifier.

MS01

Stable log message identifier.

MS02

Stable log message identifier.

NG01

Stable log message identifier.

NG02

Stable log message identifier.

NG03

Stable log message identifier.

NW01_HOST_BIND

Stable log message identifier.

NW02_PEER_CONN

Stable log message identifier.

NW03_PEER_DISC

Stable log message identifier.

NW04_NET_ERROR

Stable log message identifier.

OV01

Stable log message identifier.

OV02

Stable log message identifier.

OV03

Stable log message identifier.

P001_PULLEY_JOINT_FALLBACK

Stable log message identifier.

P002_GEAR_JOINT_FALLBACK

Stable log message identifier.

PE01

Stable log message identifier.

PE02

Stable log message identifier.

PE03

Stable log message identifier.

PE04

Stable log message identifier.

PF01_GRID_INIT

Stable log message identifier.

PF02_PATH_FOUND

Stable log message identifier.

PF03_NO_PATH

Stable log message identifier.

PG01_CELLULAR_START

Stable log message identifier.

PG02_CELLULAR_DONE

Stable log message identifier.

PL01_PIPELINE_INIT

Stable log message identifier.

PL02_STEP_ADD

Stable log message identifier.

PL03_EXEC

Stable log message identifier.

PL04_CYCLE

Stable log message identifier.

RC01

Stable log message identifier.

RL01

Stable log message identifier.

RL02

Stable log message identifier.

RL03

Stable log message identifier.

SC01_STACK_INIT

Stable log message identifier.

SC02_SCENE_PUSH

Stable log message identifier.

SC03_SCENE_POP

Stable log message identifier.

SC04_STACK_CLEAR

Stable log message identifier.

SG01

Stable log message identifier.

SG02

Stable log message identifier.

SH01_SHADER_OK

Stable log message identifier.

SH02_SHADER_ERR

Stable log message identifier.

SP01_SKEL_LOADED

Stable log message identifier.

SP02_SKEL_ANIM_MISS

Stable log message identifier.

SR01_JSON_OK

Stable log message identifier.

SR02_JSON_ERR

Stable log message identifier.

SR03_JSON_ENC

Stable log message identifier.

SR04_MSGPACK_DEC

Stable log message identifier.

SR05_MSGPACK_ENC

Stable log message identifier.

SR06_XML_OK

Stable log message identifier.

SR07_SCHEMA_PASS

Stable log message identifier.

SR08_SCHEMA_FAIL

Stable log message identifier.

SS01

Stable log message identifier.

SS02

Stable log message identifier.

SV01

Stable log message identifier.

SV01_SAVE_INIT

Stable log message identifier.

SV02

Stable log message identifier.

SV02_SAVE_WRITE

Stable log message identifier.

SV03

Stable log message identifier.

SV03_SAVE_LOAD

Stable log message identifier.

SV04

Stable log message identifier.

SV04_SAVE_ERROR

Stable log message identifier.

TH01_WORKER_INIT

Stable log message identifier.

TH02_WORKER_START

Stable log message identifier.

TH03_WORKER_DONE

Stable log message identifier.

TH04_WORKER_ERROR

Stable log message identifier.

TI01

Stable log message identifier.

TI02

Stable log message identifier.

TI03

Stable log message identifier.

TI04

Stable log message identifier.

TL01

Stable log message identifier.

TL02

Stable log message identifier.

TM01_TILEMAP_INIT

Stable log message identifier.

TM02_TILESET_ADD

Stable log message identifier.

TM03_LAYER_ADD

Stable log message identifier.

TM04_TMX_FAIL

Stable log message identifier.

TR01

Stable log message identifier.

TR02

Stable log message identifier.

TS01

Stable log message identifier.

TS02

Stable log message identifier.

TS03

Stable log message identifier.

TX01_TEX_DECODED

Stable log message identifier.

TX02_TEX_LARGE

Stable log message identifier.

TX03_TEX_ERROR

Stable log message identifier.

UP01

Stable log message identifier.

UP02

Stable log message identifier.

UP03

Stable log message identifier.

VR01

Stable log message identifier.

VR02

Stable log message identifier.

Functions

get_log_level

Return the current global log level as a static string.

set_log_level

Set the global runtime log level from a string name (off, error, warn, info, debug, trace).